Customer_Client_Reviews

+++ database/migrations/2025_06_12_073011_create_customer_reviews_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('c_user_id');
             $table->unsignedBigInteger('booking_id');
-            $table->Integer('rating');
+            $table->float('rating', 3, 1);
             $table->text('comment')->nullable();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('NO ACTION')->onUpdate('cascade');
             $table->foreign('c_user_id')->references('id')->on('users')->onDelete('NO ACTION')->onUpdate('cascade');
